Ajustes de 06/05/2023 1t

+++ database/migrations/2022_09_19_125612_create_modificacoes_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->string('NOME_TABELA')->nullable();
             $table->text('LOG')->nullable();
-            $table->integer('USER_ID')->nullable();
+            $table->bigInteger('USER_ID')->nullable();
             $table->timestamps();
         });
     }
